requirements.txt
streamlit
pandas
plotly
matplotlib
seaborn
python-dotenv
langchain
langchain-experimental
langchain-google-genai
openpyxl
tabulate
